% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIRmcmc.R
\name{SIRMCMC}
\alias{SIRMCMC}
\title{Estimation of the parameters of SIR Epidemic Model}
\usage{
SIRMCMC(I, S, m, cv, beta0, mu0)
}
\arguments{
\item{I}{numeric vector variable}

\item{S}{numeric vector variable}

\item{m}{numeric variable}

\item{beta0}{numeric variable}

\item{mu0}{numeric variable}
}
\description{
Estimation of the parameters of SIR Epidemic Model
}
